commit2

+++ WeatherApp/Extensions/KelvinToCelcius.swift
@@ -7,10 +7,10 @@
 
 import Foundation
 
-extension Float {
+extension Double {
     
     func kelvinToCeliusConverter() -> String {
-        let const : Float = 273.15
+        let const : Double = 273.15
         let kelValue = self
         let celValue = kelValue - const
         return String(format: "%.0f", celValue)
